+++ e2e/helpers/project.helper.ts
@@ -1,17 +1,17 @@
 import { projectList } from '../pages/project/list.po';
-import { projectCreate } from '../pages/project/create.po';
 import { projectView } from '../pages/project/view.po';
 import { logIn } from '../pages/login.po';
 import { Project } from '../../src/app/shared/models/project';
 import { apiTokenAdministration } from '../pages/administration/apiToken.po';
 import { permissionsAdministration } from '../pages/administration/permissions.po';
-import { projectSettingsAdministration } from '../pages/administration/projectSettings.po';
 import { User } from '../../src/app/shared/models/user';
 import { logger } from '../utils/log.util';
 import { Importer } from '../api/importer.api';
 import { EditorAPI } from '../api/editor.api';
 import usersTestData from '../data/users.json';
 import { PublicAPI } from '../api/public.api';
+import { browser } from 'protractor';
+import { UserAPI } from '../api/user.api';
 
 
 export enum PermissionType {
@@ -29,6 +29,7 @@ export class ProjectHelper {
     public importer: Importer;
     public editorAPI: EditorAPI;
     public publicAPI: PublicAPI;
+    public adminAPI: UserAPI;
     private admin = usersTestData.admin;
 
     constructor(projectName?: string) {
@@ -45,15 +46,15 @@ export class ProjectHelper {
     public async init(permissions?: { [key: string]: User; }, steps?: boolean) {
         try {
             await logIn.logInAs(this.admin.user_name, this.admin.password);
-            await this.createProject(this.project);
-            await this.openProject();
-            this.project.id = await projectView.getCurrentProjectId();
-            const token = await this.createToken(this.project);
+            const authCookie = await browser.manage().getCookie('iio78');
+            this.adminAPI = new UserAPI(decodeURIComponent(authCookie.value));
+            this.project = await this.adminAPI.createProject(this.project);
+            const token = await this.adminAPI.createToken(this.project);
             if (permissions) {
-                await this.assigneProjectPermissions(this.project, permissions);
+                await this.assigneProjectPermissions(this.project, permissions)
             }
             if (steps) {
-                await this.setSteps(true);
+                await this.adminAPI.updateProject(this.project);
             }
 
             this.importer = new Importer(this.project, token);
@@ -71,13 +72,7 @@ export class ProjectHelper {
     }
 
     public async dispose() {
-        await logIn.logInAs(this.admin.user_name, this.admin.password);
-        await projectList.isOpened();
-        await projectList.removeProject(this.project.name);
-        await logIn.menuBar.clickLogo();
-        if (await logIn.menuBar.isLogged()) {
-            await logIn.menuBar.clickLogOut();
-        }
+        await this.adminAPI.removeProject(this.project);
     }
 
     public generateBuilds = (count: number): { names: any, filenames: string[] } => {
@@ -93,50 +88,12 @@ export class ProjectHelper {
         return { names, filenames };
     }
 
-    private async setSteps(stepsState: boolean) {
-        await projectList.menuBar.administration();
-        await permissionsAdministration.sidebar.projectSettings();
-        return projectSettingsAdministration.setStepsForProject(this.project, { stepsState });
-    }
-
-    private async createProject(project: Project): Promise<void> {
-        await projectList.clickCreateProjectButton();
-        await projectCreate.fillProjectNameField(project.name);
-        await projectCreate.selectCustomer(project.customer.name);
-        return projectCreate.clickCreateButton();
-    }
-
-    private async createToken(project: Project): Promise<string> {
-        await projectList.menuBar.administration();
-        await apiTokenAdministration.sidebar.apiToken();
-        return apiTokenAdministration.generateToken(project.name);
-    }
-
     private async assigneProjectPermissions(project: Project, users: { [key: string]: User; }): Promise<void> {
-        await projectList.menuBar.administration();
-        await permissionsAdministration.sidebar.permissions();
-        await permissionsAdministration.selectProject(project.name);
         const keys = Object.keys(users);
         for (let i = 0; i < keys.length; i++) {
             const key = keys[i];
             const user: User = users[key];
-            switch (key) {
-                case PermissionType.localAdmin:
-                    await permissionsAdministration.create({ user, admin: 1, manager: 0, engineer: 0 });
-                    break;
-                case PermissionType.localEngineer:
-                    await permissionsAdministration.create({ user, admin: 0, manager: 0, engineer: 1 });
-                    break;
-                case PermissionType.localManager:
-                    await permissionsAdministration.create({ user, admin: 0, manager: 1, engineer: 0 });
-                    break;
-                case PermissionType.projectTemp:
-                case PermissionType.viewer:
-                    await permissionsAdministration.create({ user, admin: 0, manager: 0, engineer: 0 });
-                    break;
-                default:
-                    logger.info(`Local Permissions for ${key} are not required`);
-            }
+            await this.adminAPI.assigneProjectPermission(this.project, user, key as PermissionType);
         }
     }
 }