+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## 0.3.10 (unreleased)
 
 Features:
+  - Update Tests to use Admin API to create preconditions -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/100)
 
 Bugfixes:
   - Test run View Page performance is bad -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/99)
+  - API Token page: Link to Aquality Tracking API leads to null -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/84)
 
 ## 0.3.9 (2020-04-21)